Initial

+++ src/queries/engineering-service.ts
@@ -1,6 +1,7 @@
 "use server";
 
 import { connect } from "@/db/database-config";
+import EngineeringCategory from "@/models/engineering-category.model";
 import EngineeringService from "@/models/engineering-service.model";
 import { ServiceData } from "@/types/service-data";
 
@@ -10,9 +11,9 @@ export const getEngineeringServiceList = async () => {
   try {
     const engineeringServiceList = await EngineeringService.find({}).populate({
       path: "category",
-      select: "title",
+      select: "title category",
     });
-    return JSON.stringify(engineeringServiceList);
+    return JSON.parse(JSON.stringify(engineeringServiceList));
   } catch (error: any) {
     throw new Error(error.message || "Failed to fetch service list");
   }
@@ -23,8 +24,7 @@ export const getEngineeringServiceById = async (serviceId: string) => {
 
   try {
     const engineeringServiceById = await EngineeringService.findById(serviceId);
-
-    return JSON.stringify(engineeringServiceById);
+    return JSON.parse(JSON.stringify(engineeringServiceById));
   } catch (error: any) {
     throw new Error(error.message || "Failed to fetch get category details");
   }
@@ -57,7 +57,19 @@ export const createEngineeringService = async (serviceData: ServiceData) => {
         },
       },
     });
-    return JSON.stringify(engineeringServiceDetails);
+
+    const categoryId = serviceData.category;
+    let category = await EngineeringCategory.findById(categoryId);
+    if (!category) {
+      throw new Error("Category not found!");
+    }
+
+    if (!category.services.includes(engineeringServiceDetails._id)) {
+      category.services.push(engineeringServiceDetails._id);
+      await category.save();
+    }
+
+    return JSON.parse(JSON.stringify(engineeringServiceDetails));
   } catch (error: any) {
     throw new Error(error.message || "Failed to create service");
   }
@@ -74,7 +86,19 @@ export const updateEngineeringService = async (
       await EngineeringService.findByIdAndUpdate(serviceId, serviceData, {
         new: true,
       });
-    return JSON.stringify(engineeringServiceDetails);
+
+    const updatedCategoryId = serviceData.category;
+    let category = await EngineeringCategory.findById(updatedCategoryId);
+    if (!category) {
+      throw new Error("Category not found!");
+    }
+
+    if (!category.services.includes(serviceId)) {
+      category.services.push(serviceId);
+      await category.save();
+    }
+
+    return JSON.parse(JSON.stringify(engineeringServiceDetails));
   } catch (error: any) {
     throw new Error(error.message || "Failed to update service");
   }
@@ -84,21 +108,56 @@ export const deleteEngineeringService = async (serviceId: string) => {
   await connect();
 
   try {
-    const engineeringServiceDetails =
-      await EngineeringService.findByIdAndDelete(serviceId);
-    return JSON.stringify(engineeringServiceDetails);
+    const service = await EngineeringService.findById(serviceId);
+    if (!service) {
+      throw new Error("Service not found!");
+    }
+
+    const categoryId = service.category;
+
+    const category = await EngineeringCategory.findById(categoryId);
+    if (!category) {
+      throw new Error("Category not found!");
+    }
+
+    category.services = category.services.filter(
+      (_id: string) => _id.toString() !== serviceId
+    );
+
+    await category.save();
+    await EngineeringService.findByIdAndDelete(serviceId);
   } catch (error: any) {
     throw new Error(error.message || "Failed to delete service");
   }
 };
 
-export const deleteEngineeringServiceList = async () => {
+export const deleteEngineeringServiceList = async ()=> {
   await connect();
 
   try {
-    return await EngineeringService.deleteMany({});
+    const services = await EngineeringService.find({});
+    const categoryIds: Set<string> = new Set(
+      services.map((service) => service.category.toString())
+    );
+
+    await EngineeringService.deleteMany({});
+
+    for (const categoryId of categoryIds) {
+      const category = await EngineeringCategory.findById(categoryId);
+      if (category) {
+        category.services = category.services.filter(
+          (serviceId: string) =>
+            !services.some(
+              (service) => service._id.toString() === serviceId.toString()
+            )
+        );
+        await category.save();
+      }
+    }
+
+    return { message: "All services deleted and categories updated." };
   } catch (error: any) {
-    throw new Error(error.message || "Failed to delete service");
+    throw new Error(error.message || "Failed to delete services");
   }
 };
